+++ test/test_models/test_seq_lab_net.py
@@ -63,6 +63,14 @@ def test_ecg_seq_lab_net():
     assert isinstance(doi, list)
     assert all([isinstance(d, str) for d in doi]), doi
 
+    with pytest.raises(RuntimeError, match="Maybe you are trying to load a model trained with numpy 1"):
+        ECG_SEQ_LAB_NET.from_remote(
+            url="https://drive.google.com/uc?id=18Jta73DjqXVarEYjN_CWeYqM8rH7I3An",
+            model_dir=_TMP_DIR,
+            filename="best_model.pth",
+            weights_only="auto",
+        )
+
 
 def test_warns_errors():
     inp = torch.randn(2, 12, 2000)